Have finished a solid portion of the foundation work
learned more then i wanted to with how sqlight deals with different characters in different scanrios
added the second primary table.
Things going to plan, any new tables only need to know of and link to the second table.
When the Primary FileGui from the first table or the timechain in the second table gets deleted, everything linked should get purged. To do with the Timechain.

That or ive misinterprited the order im supposed to set foreign keys and everything will blow up yay!
Another thing to test tomorrow.